fixes #112
changing the ERROR log statements to WARN in PluginManager, as they do not
really represent errors, but just inform of potential threats, that
lie outside of our responsibility.

checking for plugin implementation upon source adaptor creation in
the adaptor manager in order to prevent the creation of new source adaptors
without implementation.

changing some of the test data to reflect the newest c3po adaptor version.