resolving a bug in the pronom adaptor
after the first complete crawl of the pronom
adaptor the hasNext() method had a problem
and always returned false after the first
batch of the second crawl. This resulted
in starvation and it was not checked, whether
new results were appended at the end.
Now the whole crawl is done before hasNext()
returns false to make sure that there were
no new records appended to PRONOM