fixes #31
adding a policy model helper that
can read and load policy model files
conforming to the SCAPE policy schema.

The current front end allows upload
of one or multiple files. The uploaded
files are stored in the .scout folder
and are loaded into the knowledge base
under a special named model.

Also browsing of the loaded policies is possible
both through web/browse and web/browse/policy.

Further management, such as delete and
is not yet supported.